fix invoice detail

+++ lib/models/invoice_detail_dto.dart
@@ -8,7 +8,7 @@ class InvoiceDetailDTO {
   int? status;
   int? vatAmount;
   int? amount;
-  int? vat;
+  double? vat;
   String? bankId;
   String? bankAccount;
   String? bankShortName;
